requirements.txt
google-adk
beautifulsoup4
requests
python-dotenv
openai
